Created delete product from basket route for basket page in userRouter.ts, created deleteProductFromBasketController. Edited getProductFromBasket and deleteProductFromBasket functions in basket slice (db). Created get products from basket and orders route for basket page in userRouter.ts, created getProductsFromBasketController and getFullPathToImage function in product slice (db)

+++ src/db/slices/basket.ts
@@ -1,7 +1,7 @@
 import db from "../db.js";
 import { FieldPacket, RowDataPacket } from "mysql2/promise";
 import { IBasketProduct } from "../../lib/types.js";
-import { getProduct } from "./products.js";
+import { getProduct, getFullPathToImage } from "./products.js";
 
 export async function checkExistOfProductInBasket(productId: number, userId: number): Promise<boolean> {
   const data: [(RowDataPacket & { product_id: number })[], FieldPacket[]] = await db.query(
@@ -15,18 +15,22 @@ export async function checkExistOfProductInBasket(productId: number, userId: num
   }
 }
 
-export async function getProductOfBasket(productId: number): Promise<{ product: IBasketProduct } | null> {
+export async function getProductOfBasket(productId: number, userId: number): Promise<{ product: IBasketProduct } | null> {
   const data: [(RowDataPacket & IBasketProduct)[], FieldPacket[]] = await db.query(
     `
-        SELECT p.id AS id, p.title AS title, p.image AS image, b.quantity AS quantity, p.price AS price FROM products AS p
+        SELECT p.id AS id, p.title AS title, p.image AS image, b.quantity AS orderedQuantity, p.quantity AS availableQuantity, p.price AS price FROM products AS p
         LEFT JOIN basket AS b ON p.id = b.product_id
-        WHERE p.id = "${productId}"
+        WHERE p.id = "${productId}" AND b.user_id = "${userId}"
     `
   );
 
   if (data[0][0]) {
+    const product = data[0][0];
+    // коректируем путь к изображению
+    product.image = getFullPathToImage(product.image);
+
     return {
-      product: data[0][0]
+      product
     };
   } else {
     return null;
@@ -36,13 +40,18 @@ export async function getProductOfBasket(productId: number): Promise<{ product: 
 export async function getProductsFromBasket(userId: number): Promise<IBasketProduct[]> {
   const data: [(RowDataPacket & IBasketProduct)[], FieldPacket[]] = await db.query(
     `
-        SELECT p.id AS id, p.title AS title, p.image AS image, b.quantity AS quantity, p.price AS price FROM products AS p
+        SELECT p.id AS id, p.title AS title, p.image AS image, b.quantity AS orderedQuantity, p.quantity AS availableQuantity, p.price AS price FROM products AS p
         LEFT JOIN basket AS b ON p.id = b.product_id
         WHERE b.user_id = "${userId}"
     `
   );
 
-  return data[0];
+  const products = data[0].map(el => {
+    el.image = getFullPathToImage(el.image);
+    return el;
+  });
+
+  return products;
 }
 
 export async function addProductToBasket(productId: number, userId: number): Promise<{ product: IBasketProduct } | null> {
@@ -53,7 +62,7 @@ export async function addProductToBasket(productId: number, userId: number): Pro
     `
   );
 
-  const product = await getProductOfBasket(productId);
+  const product = await getProductOfBasket(productId, userId);
   return product;
 }
 
@@ -69,7 +78,7 @@ export async function updateBasketProductQuantity(productId: number, quantity: n
         `UPDATE basket SET quantity = "${quantity}" 
         WHERE product_id = "${productId}" AND user_id = "${userId}"`
       );
-      const product = await getProductOfBasket(productId);
+      const product = await getProductOfBasket(productId, userId);
       return product;
     }
   }
@@ -77,8 +86,8 @@ export async function updateBasketProductQuantity(productId: number, quantity: n
   return null;
 }
 
-export async function deleteProductFromBasket(productId: number): Promise<{ product: IBasketProduct } | null> {
-  const product = await getProductOfBasket(productId);
-  await db.query(`DELETE FROM basket WHERE product_id = "${productId}"`);
+export async function deleteProductFromBasket(productId: number, userId: number): Promise<{ product: IBasketProduct } | null> {
+  const product = await getProductOfBasket(productId, userId);
+  await db.query(`DELETE FROM basket WHERE product_id = "${productId}" AND user_id = "${userId}"`);
   return product;
 }
